+++ packages/locale/src/index.ts
@@ -2,7 +2,7 @@
  * Tencent is pleased to support the open source community by making
  * 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) available.
  *
- * Copyright (C) 2021 THL A29 Limited, a Tencent company.  All rights reserved.
+ * Copyright (C) 2025 Tencent.  All rights reserved.
  *
  * 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) is licensed under the MIT License.
  *
@@ -27,4 +27,5 @@
 export { default as en } from './lang/en';
 export type { Language } from './lang/en';
 
+export { default as jp } from './lang/ja-jp';
 export { default as zhCn } from './lang/zh-cn';