+++ packages/modal/src/modal.tsx
@@ -2,7 +2,7 @@
  * Tencent is pleased to support the open source community by making
  * 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) available.
  *
- * Copyright (C) 2021 THL A29 Limited, a Tencent company.  All rights reserved.
+ * Copyright (C) 2025 Tencent.  All rights reserved.
  *
  * 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) is licensed under the MIT License.
  *
@@ -76,6 +76,15 @@ export default defineComponent({
         top: props.top,
       };
     });
+
+    watch(
+      () => [props.zIndex],
+      () => {
+        if (/\d+/.test(`${props.zIndex}`)) {
+          zIndex.value = props.zIndex;
+        }
+      },
+    );
 
     watch(
       () => props.isShow,