+++ packages/tab/src/props.ts
@@ -2,7 +2,7 @@
  * Tencent is pleased to support the open source community by making
  * 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) available.
  *
- * Copyright (C) 2021 THL A29 Limited, a Tencent company.  All rights reserved.
+ * Copyright (C) 2025 Tencent.  All rights reserved.
  *
  * 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) is licensed under the MIT License.
  *
@@ -128,6 +128,9 @@ export const tabProps = {
   validateActive: PropTypes.bool.def(true),
   activeBarSize: PropTypes.number.def(2),
   activeBarColor: PropTypes.string.def('#3a84ff'),
+  beforeChange: {
+    type: Function,
+  },
 };
 
 export const tabNavProps = {