+++ scripts/cli/utils/generate-deration.ts
@@ -2,7 +2,7 @@
  * Tencent is pleased to support the open source community by making
  * 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) available.
  *
- * Copyright (C) 2021 THL A29 Limited, a Tencent company.  All rights reserved.
+ * Copyright (C) 2025 Tencent.  All rights reserved.
  *
  * 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) is licensed under the MIT License.
  *
@@ -23,10 +23,12 @@
  * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
  * IN THE SOFTWARE.
  */
-import { exec } from 'child_process';
+import { execFile } from 'child_process';
 import path from 'path';
 import { promisify } from 'util';
 
 export default async () => {
-  await promisify(exec)(`tsc -p ${path.resolve(__dirname, '../tsconfig.declaration.json')}`);
+  // await promisify(exec)(`tsc -p ${path.resolve(__dirname, '../tsconfig.declaration.json')}`);
+  const tscPath = path.resolve(__dirname, '../tsconfig.declaration.json');
+  await promisify(execFile)('tsc', ['-p', tscPath]);
 };