+++ site/views/dialog/base-demo.vue
@@ -5,14 +5,33 @@
       text
       @click="handleShow"
     >
-      打开 dialog
+      打开 dialog1
     </bk-button>
     <bk-dialog
       v-model:is-show="isShow"
-      title="描述"
+      :draggable="true"
+      :show-mask="false"
+      title="描述1"
       quick-close
     >
-      <div>hello world!</div>
+      <div>hello world - 111!</div>
+    </bk-dialog>
+
+    <bk-button
+      theme="primary"
+      text
+      @click="handleShow2"
+    >
+      打开 dialog2
+    </bk-button>
+    <bk-dialog
+      v-model:is-show="isShow2"
+      :draggable="true"
+      :show-mask="false"
+      title="描述2"
+      quick-close
+    >
+      <div>hello world - 111!</div>
     </bk-dialog>
   </div>
 </template>
@@ -21,8 +40,13 @@
   import { ref } from 'vue';
 
   const isShow = ref(false);
+  const isShow2 = ref(false);
 
   const handleShow = () => {
     isShow.value = true;
+  };
+
+  const handleShow2 = () => {
+    isShow2.value = true;
   };
 </script>